fixed the help regular expression, change to global prefix
Global prefixes are stripped of whitespace during checks, so now a
prefix of e.g. 'robot: ' doesn't *require* a space in a message line to
match. The space is still shown in help text.

The help regular expression fix was to make sure the help text of a
command that is a substring of another command doesn't superscede the
other command. e.g. the commands 'reload' and 'reloadall', trying to run
reloadall would just display the help for reload.