changed plugin reloading to module reloading
Since python reloads code on the module level, it causes problems if you
reload a module but continue to use instances from classes made from the
old module. The module globals get updated, so when you call things like
super(), it references the new definition of the class and raises a
TypeError.
Solution: always reload all plugins in a module.