Added a different denied msg for commands when users aren't identified
Now, when a user is not identified, and they try to issue a command, it
also checks to see if the user /would/ have had permission were they
identified under the authname of the nick they're using currently. If
so, delivers a more helpful message. The implementation is a bit of a
hack imo but it should get the job done.