fixed bug due to events going to unloaded plugins
if an event triggers a plugin to be unloaded or reloaded, and the plugin
is listening for that same event, then it is possible for the plugin to
get the event after it has been unloaded or reloaded. This change in the
transport layer checks to see if a plugin has been unloaded before
sending it an event.

Note that the plugin *could* still get an event that also triggers its
own unloading. It depends on the iteration order of the plugins. This
change only ensures that if it *does* get called, the plugin has not
(yet) been unload