added exception handlers around plugin calls
If a plugin handler raises an exception, it would propagate all the way
back to twisted and prevent the transport layer from calling any other
parameters. In the worst case, this prevented me from running a !plugin
reload because the buggy plugin was erroring before the plugin reloader
was getting that line.

Now all exceptions are caught at the transport layer and ignored, after
printing a traceback to the console.