multiple changes to VOTD plugin
Added a scalefactor parameter, which changes how entries are counted. It
changes how the numbers are rounded. Higher numbers will make for
smoother probability curves. Or something like that.

Added a chance section to the config, which is purely informational for
external programs that may want to read the config to look at the
current status. As kind of an api.

Fixed a couple bugs and potential bugs. multipliers dict in config is
now properly pruned. Delayed drawings will be canceled if the plugin is
unloaded.

Users' entries are no longer reset on win. Only the multiplier gets
adjusted for the winner.