admin plugin and ircop plugin changes
ircop plugin now exports ircop.ban and ircop.unban commands, for
consistency with other commands even though there isn't a connector
interface for them.

admin plugin modified to work with the new ircop.ban and ircop.unban.

This change made it easier to use the proper interface for the admin
quiet and unquiet commands, which DO have a connector interface, but
share code with the ban and unban commands.