admin plugin changes
added a !mode command. mode command takes an optional parameter but
doesn't check to see if the mode actually takes a parameter. so it is
possible for a malformed mode to disrupt other modes that get sent in
the same message to the server.

changed how _do_moderequest() works to return after the whois completes,
and do the op request in parallel (but still callback for errors).

ircop stealth change: changed grace period for buffer additions to 0.2
seconds from 0.5 seconds. Reason: the time is dwarfed by the time it
takes to gain op from chanserv, so we usually have 2 seconds or so grace
period anyways. This change makes commands when holding-op go through
faster.