ircop changes, comments, error handling...
Updated the comments and docstrings a bit for accuracy

Added proper error handling around connector operations that happen
because an OP request failed.

_wait_for_op() calls that happen in the request handlers are
optimizations to acquire op while the buffer timer is waiting for any
other possible requests. (OP will still be requested when the buffer is
processed, and errors there are not ignored) For this reason, errors are
silenced by setting a noop error handler. This commit changes the error
handler to only silence OpFailed exceptions. Others will be propagated
and dumped to the console when the deferred is garbage collected.