added mode parameter error checking
in the irc plugin, added a request to return modes that take a
parameter.

in the ircop plugin, checks that a mode request correctly does or does
not have a parameter, and raises valueerror if it's incorrect.

in the admin plugin, catches valueerror and replies with an error to the
user.