misc changes to Owner plugin and wotd plugin
added !code as an alias for !source. Or the other way around, I forget.

Changed so the bot doesn't kick if you say the word of the day as the
entire message, if it's a repeat of someone else. The kick on repeat
was just meant to catch people saying the entire line that happened to
accidentially trigger it.