deleted docs, added TODO section about the rewrite
I wanted to remove the docs because they were becoming more and more
out of date. Even though they still had some relevant information, I
think the code and the docstrings speak for themselves. Besides, this
isn't some project with a huge userbase, if someone is confused they can
just ask me.

I added a section to the TODO document about my rewrite ideas, and the
goals for the rewrite. I plan to add more details to that section soon.