fixed OP behavior details, minor bug
When the bot has OP from human intervention (as opposed to OP gained in
the process of submitting OP-requiring commands), it should not
automatically drop OP if it does happen to submit some commands.

A long-standing bug meant that if the become_op request was called, it
would ignore existing OP status and remove OP, ignoring the intended
behavior. This commit adds a simple check to see if the bot already has
OP when a become_op request is submitted, and if so, ignores it.

Also updated a comment with some notes to make some implementation
details more clear.