comments and delay to setting mode timers upon startup
Mode request timers will now wait 30 seconds upon bot startup before
taking any action. This is to prevent modes being sent to the server
before we're connected. It's a bit of a hack though. see comment in code
for more details.

Also touched up and reworded a few comments and docstrings for clarity.