fixed bug in spam plugin
If action is to be taken, it clears out the lines as a hackish way to
not take action twice if the user continues to flood between when the
plugin decides to take action and when action is actually taken.

Turns out, clearing the lines *after* action is taken is not a very
useful way of accomplishing this.