fixed bug in admin.py involving duplicate timer events
If the plugin was reloaded within 30 seconds of starting, the timer that
calls _set_all_timers wasn't canceled, causing the unloaded plugin to
set timers and causing duplicate events to be processed.

Now the initial _set_all_timers timer is properly canceled when the
plugin is unloaded.

Also the later timers are properly deleted after they are canceled in
_set_all_timers, avoiding some AlreadyCancelled errors.