Changes to time parsing in admin commands
when you do a !ban or a !quiet or a !unban or a !unquiet or a !mode with
a time/duration, it will now say or PM you the time it parsed so you
know it got it right. also !quiet with a message that's not a duration
no longer errors but chooses the default time.