add realname
The "realname" field of an IRC client is accessible when somebody runs a
whois on the client in question, although only rarely is it used for
people's actual real names. This commit adds a realname configuration
option.

This option is populated during the configuration initialisation wizard
with a sensible default for people who don't bother to customise things:
a string pointing to the bot's operator, which is a nice courtesy to
users (or IRC operators) who want to contact the person in charge.

The configuration can be altered manually by editing the "realname"
parameter of the irc.IRCBotPlugin.json file in the configuration
directory. One might wish to link to the bot's source code, for example.